QR_code_generator/backend/requirements.txt
Flask
Flask-Cors
python-dotenv
qrcode[pil]
huggingface_hub
requests
